unixfd: Fix assertion failure constructing UnixFd for invalid fd
File descriptors are represented as the C int type, but only non-negative
values represent a valid fd, with all negative values representing the
absence of a fd (a NULL-like value).

Previously, make_fd() accepted negative fds, but then UnixFd_tp_new
would crash with an assertion failure. Instead, range-check the value
according to the semantically valid range. If it isn't, raise the same
error we previously raised for values that don't fit in an int.

Signed-off-by: Simon McVittie <smcv@collabora.com>