connection: Set NO_REPLY flag if there's no handler
If we do an async call with no handler, it doesn't matter whether the
implementation replies with a result or an error, because either way
we will ignore it. Normally, the method implementation is expected to
reply to all method calls, but with the NO_REPLY flag set, it can
optimize by not sending a reply at all, which also means the message
bus does not have to keep track of pending replies.

Co-authored-by: Simon McVittie <smcv@collabora.com>